+++ validator/metrics.go
@@ -1,0 +1,27 @@
+package validator
+
+import (
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
+	"log"
+)
+
+var (
+	metricsRunningIBFTsCount = promauto.NewGauge(prometheus.GaugeOpts{
+		Name: "ssv:validator:running_ibfts_count_all",
+		Help: "Count all running IBFTs",
+	})
+	metricsRunningIBFTs = promauto.NewGaugeVec(prometheus.GaugeOpts{
+		Name: "ssv:validator:running_ibfts_count",
+		Help: "Count running IBFTs by validator pub key",
+	}, []string{"pubKey"})
+)
+
+func init() {
+	if err := prometheus.Register(metricsRunningIBFTsCount); err != nil {
+		log.Println("could not register prometheus collector")
+	}
+	if err := prometheus.Register(metricsRunningIBFTs); err != nil {
+		log.Println("could not register prometheus collector")
+	}
+}
